tester: UI float display, timer/cleanup fixes; npifc: resource cleanup and robust registry path; tester UI: add DLL/App signature and version fields; unify tir_data typedef; linuxtrack: robust lib discovery (USERPROFILE fallback, distro paths); NPClient: stop auto TrackIR launch/suspend; wake/request_frames/notification + wait loop; client/rest/check_data: fix formats and buffer sizes; update npview.rc.in template; add hardening plan doc

+++ src/wine_bridge/client/rest.c
@@ -278,7 +278,7 @@ bool runFile(const char *file)
 printf("Going to run this: %s\n", q_exe);
   bool res = CreateProcess(NULL, q_exe, NULL, NULL, false, NORMAL_PRIORITY_CLASS, NULL, NULL, &si, &pi);
   if(!res){
-    printf("Failed! (%d)\n", GetLastError());
+    printf("Failed! (%ld)\n", (long)GetLastError());
   }
   CloseHandle(pi.hProcess);
   CloseHandle(pi.hThread);
